+++ R/hist.R
@@ -76,7 +76,7 @@ setMethod("hist", signature(x="SpatRaster"),
 		v <- stats::na.omit(values(x))
 	} else {
 		# TO DO: make a function that does this by block and combines  all data into a single histogram
-		v <- spatSample(x, maxcell, method="regular", as.df=FALSE, as.raster=FALSE)
+		v <- spatSample(x, maxcell, method="regular", as.df=FALSE, as.raster=FALSE, warn=FALSE)
 		msg <- paste("a sample of", round(100 * length(v) / ncell(x)), "% of the cells was used", sep="")
 		if (any(is.na(v))) {
 			v <- stats::na.omit(v)