+++ R/plot2rasters.R
@@ -56,7 +56,7 @@
 	}
 
 
-	x <- as.matrix(spatSample(c(x,y), size=maxcell, method="regular", as.raster=FALSE))
+	x <- as.matrix(spatSample(c(x,y), size=maxcell, method="regular", as.raster=FALSE, warn=FALSE))
 	# y <- as.matrix(spatSample(y, size=maxcell, method="regular", as.raster=FALSE))
 
 	y <- x[,c((nlx+1):ncol(x))]