+++ R/plotRGB.R
@@ -59,7 +59,7 @@ function(x, r=1, g=2, b=3, a=NULL, scale, maxcell=500000, mar=0, stretch=NULL, e
 	if (!is.null(ext)) {
 		x <- crop(x, ext)
 	}
-	x <- spatSample(x, maxcell, method="regular", as.raster=TRUE)
+	x <- spatSample(x, maxcell, method="regular", as.raster=TRUE, warn=FALSE)
 	if (!is.null(zlim)) {
 		if (length(zlim) == 2) {
 			zlim <- sort(zlim)